MAGE-1154: add RecordBuilder for Products
